feat: Add modifierCollecteByAdmin method to CollecteService
This commit adds a new method, modifierCollecteByAdmin, to the CollecteService. This method allows an admin user to modify a collecte by updating its data in the Firestore database. The method takes a Collecte object as input and updates the corresponding document in the 'Collecte' collection.

The method first creates an updatedDonateurData object by spreading the properties of the input collecte object. Then, it retrieves the collecte document reference using the collecte's id and updates the document with the updatedDonateurData.

This new method provides a way for admin users to modify collecte data, expanding the functionality of the CollecteService.