chore: Update proxy configuration for local development
This commit modifies the `package.json` file to update the `start` script command. The `ng serve` command now includes the `--proxy-config` option with the value `proxy.conf.json`. This configuration enables proxying requests to a different server during local development, improving the development experience when working with backend APIs.

This change ensures that requests made to the `/api` path are proxied to the specified server defined in the `proxy.conf.json` file.